formulario para agregar coordenadas de notificacion de huecos

+++ geomatica/core/models.py
@@ -4,8 +4,8 @@ from django.db import models
 
 class Hueco(models.Model):
     direccion = models.CharField(max_length=200)
-    latitud = models.IntegerField(default=0)
-    longitud = models.IntegerField(default=0)
+    latitud = models.FloatField(default=0.0)
+    longitud = models.FloatField(default=0.0)
     foto = models.ImageField(upload_to = 'static/image_user/')
 
 class Trafico(models.Model):
